[mod] alarm dialog navigation 변경
