+++ app/src/main/java/com/dgu/smartcareapp/presentation/CreationTodo/CreateTodoNavigation.kt
@@ -2,6 +2,7 @@ package com.dgu.smartcareapp.presentation.CreationTodo
 
 import android.content.Context
 import androidx.compose.ui.Modifier
+import androidx.lifecycle.LifecycleCoroutineScope
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
@@ -15,6 +16,7 @@ fun NavGraphBuilder.todoNavGraph(
 //    onButtonClick: (time: String, toDoTitle: String) -> Unit = { _, _ -> },
     onButtonClick: () -> Unit = {},
     context: Context,
+    lifecycleScope: LifecycleCoroutineScope,
     onNavigationIconClick: () -> Unit = {},
 ) {
     composable(route = ToDo.ROUTE) {
@@ -22,6 +24,7 @@ fun NavGraphBuilder.todoNavGraph(
             modifier = modifier,
             onButtonClick = onButtonClick,
             context = context,
+            lifecycleScope = lifecycleScope,
             onNavigationIconClick = onNavigationIconClick
         )
     }
