+++ app/src/main/java/com/dgu/smartcareapp/presentation/main/MainActivity.kt
@@ -2,16 +2,28 @@ package com.dgu.smartcareapp.presentation.main
 
 import android.content.Context
 import android.os.Bundle
+import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleEventObserver
+import androidx.lifecycle.compose.LocalLifecycleOwner
+import androidx.lifecycle.lifecycleScope
+import com.dgu.smartcareapp.alarm.AlarmManager
+import com.dgu.smartcareapp.component.AlarmDialog
 import com.dgu.smartcareapp.ui.theme.SmartCareAppTheme
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.flow.collectLatest
+import kotlinx.coroutines.launch
 
 val LocalDeviceSizeComposition = staticCompositionLocalOf {
     com.dgu.smartcareapp.presentation.main.DeviceSize.MEDIUM
@@ -19,13 +31,50 @@ val LocalDeviceSizeComposition = staticCompositionLocalOf {
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
+
+    // todo rememberSaveble로 수정
+    val isShow = mutableStateOf(false)
+    val toDoTitle: MutableState<String?> = mutableStateOf(null)
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             val navigator: MainNavigator = rememberMainNavigator()
             val deviceWidth = applicationContext?.resources?.displayMetrics?.widthPixels ?: 0
 
+            val lifecycleOwner = LocalLifecycleOwner.current
+
+            // 이 부분 확인해보기
+            DisposableEffect(lifecycleOwner) {
+                val observer = LifecycleEventObserver { _, event ->
+                    if (event == Lifecycle.Event.ON_RESUME) {
+                        lifecycleScope.launch {
+                            // 가장 최근 알람을 받아 옴
+                            AlarmManager.alarm.collectLatest {
+                                Log.d("dana", "알람왔다!")
+                                isShow.value = true
+                                toDoTitle.value = it
+                            }
+                        }
+                    }
+                }
+                lifecycleOwner.lifecycle.addObserver(observer)
+                onDispose {
+                    lifecycleOwner.lifecycle.removeObserver(observer)
+                }
+            }
+
             SmartCareAppTheme {
+
+                if (isShow.value) {
+                    toDoTitle.value?.let {
+                        AlarmDialog(
+                            toDoTitle = it,
+                            onConfirm = {isShow.value = false}
+                        )
+                    }
+                }
+
                 Surface(
                     modifier = Modifier.fillMaxSize(),
                     color = Color.White
@@ -35,7 +84,7 @@ class MainActivity : ComponentActivity() {
                             deviceWidth
                         )
                     ) {
-                        MainScreen(navigator, this)
+                        MainScreen(navigator, this, lifecycleScope)
                     }
                 }
             }
